Update README.md
The objective of this project is to build a predictive model using linear regression to estimate house prices based on a dataset with relevant features. Linear regression is a fundamental machine learning algorithm that allows for the prediction of numerical outcomes, and this project offers hands-on experience in developing, evaluating, and interpreting such a model.